Code refactored & code commented
